+++ django_enum/tests/converters/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class Converters(AppConfig):
+    name = 'django_enum.tests.converters'
+    label = name.replace('.', '_')