+++ django_enum/tests/tmpls/templatetags/test_tags.py
@@ -1,5 +1,6 @@
-from django import template
 from enum import Enum
+
+from django import template
 
 register = template.Library()