+++ django_enum/tests/urls.py
@@ -4,7 +4,8 @@ from django.urls import include, path
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('djenum/', include('django_enum.tests.djenum.urls'))
+    path('djenum/', include('django_enum.tests.djenum.urls')),
+    path('', include('django_enum.tests.converters.urls'))
 ]
 
 if 'django_enum.tests.enum_prop' in settings.INSTALLED_APPS:  # pragma: no cover